feat: Add backend-defined data refresh to V2 apps
